+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # v3.14.2
-* Dnd5e: Temporary workaround for vision-5e hooks while I work on properly hooking up Stealthy to it.
-* Dnd5e: Updated pt-BR translations (Thanks Kharmans)
+* Dnd5e: If vision-5e is enabled, add perception vs stealth checks to the following detection modes: `devilsSight`, `etherealSight`, `hearing`, `lightPerception`, `seeAll`, `seeInvisibility`, and `witchSight`
+* Dnd5e: Updated pt-BR translations (thanks Kharmans)
 * Dnd5e: Added both name and label fields to the language .json files for Dim/Dark
 * PF1: Disabled Stealthy until I can figure out a better way to capture skill rolls
 * Dnd4e: Disabled Stealthy until I can figure out a better way to capture skill rolls